requirements.txt
streamlit
openai
transformers
torch
PyMuPDF
python-dotenv